fix: delete, update에 대해 admin이거나 해당 유저만 가능하도록 설정

+++ issue/src/main/java/com/sweprj/issue/service/CommentService.java
@@ -1,6 +1,7 @@
 package com.sweprj.issue.service;
 
 import com.sweprj.issue.DTO.CommentDTO;
+import com.sweprj.issue.config.jwt.JwtTokenProvider;
 import com.sweprj.issue.domain.Comment;
 import com.sweprj.issue.domain.Issue;
 import com.sweprj.issue.domain.User;
@@ -8,6 +9,8 @@ import com.sweprj.issue.exception.ResourceNotFoundException;
 import com.sweprj.issue.repository.CommentRepository;
 import com.sweprj.issue.repository.IssueRepository;
 import com.sweprj.issue.repository.UserRepository;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.text.SimpleDateFormat;
@@ -19,11 +22,13 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final IssueRepository issueRepository;
     private final UserRepository userRepository;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public CommentService(CommentRepository commentRepository, IssueRepository issueRepository, UserRepository userRepository) {
+    public CommentService(CommentRepository commentRepository, IssueRepository issueRepository, UserRepository userRepository, JwtTokenProvider jwtTokenProvider) {
         this.commentRepository = commentRepository;
         this.issueRepository = issueRepository;
         this.userRepository = userRepository;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
     public CommentDTO createComment(Long issueId, Long userId, String content) {
@@ -48,6 +53,7 @@ public class CommentService {
     public CommentDTO updateComment(Long commentId, String content) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new ResourceNotFoundException("Comment not found with id " + commentId));
+        checkIfAuthorized(comment);
         if (content != null) {
             comment.setContent(content);
         }
@@ -57,6 +63,7 @@ public class CommentService {
 
     public void deleteComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new ResourceNotFoundException("Comment not found"));
+        checkIfAuthorized(comment);
         commentRepository.delete(comment);
     }
 
@@ -68,5 +75,15 @@ public class CommentService {
         commentDTO.setCommentedAt(new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ").format(comment.getCommentedAt()));
         commentDTO.setIssueId(comment.getIssue().getId());
         return commentDTO;
+    }
+
+    private void checkIfAuthorized(Comment comment) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String token = authentication.getCredentials().toString();
+        Long userId = jwtTokenProvider.getUserFromJwt(token);
+        String role = jwtTokenProvider.getRoleFromJwt(token);
+        if (!role.equals("ADMIN") && !comment.getCommenter().getUserId().equals(userId)) {
+            throw new ResourceNotFoundException("You are not authorized to perform this action");
+        }
     }
 }
